+++ app/src/js/components/DataUpload/overview.js
@@ -1,152 +1,228 @@
 'use strict';
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import _config from '../../config';
 import { loadToken } from '../../utils/auth';
+import Loading from '../LoadingIndicator/loading-indicator';
 import localUpload from 'edpub-data-upload-utility';
-import { listFileUploadsBySubmission } from '../../actions';
+import { listFileUploadsBySubmission, listFileDownloadsByKey } from '../../actions';
+import { shortDateShortTimeYearFirstJustValue, storage } from '../../utils/format';
+import Table from '../SortableTable/SortableTable';
 
 class UploadOverview extends React.Component {
   constructor() {
     super();
-    this.state = { loaded: false, files: '', hiddenFileInput: React.createRef(null), statusMsg: 'Select a file', uploadFile: '' };
-    /* const [statusMsg, setStatusMsg] = useState('Select a file');
-    const [uploadFile, setUploadFile] = useState('');
-    const [submissionId, setSubmissionId] = useState(''); */
-    this.updateFileList = this.updateFileList.bind(this);
+    this.state = { 
+      loaded: false, 
+      hiddenFileInput: React.createRef(null), 
+      statusMsg: 'Select a file', 
+      uploadFile: '', 
+      keys: []
+    };
     this.handleClick = this.handleClick.bind(this);
     this.handleChange = this.handleChange.bind(this);
+    this.getFileList = this.getFileList.bind(this);
+    this.validateFile = this.validateFile.bind(this);
+    this.keyLookup = this.keyLookup.bind(this);
+    this.resetInputWithTimeout = this.resetInputWithTimeout.bind(this);
   }
-
-  componentDidUpdate() {
-    console.log('componentDidUpdate called')
-    this.updateFileList()
-  }
-
-  updateFileList(){
-    const { dispatch } = this.props;
-    let submissionId = '';
-    if (window.location.href.indexOf('requests/id') >= 0) {
-      submissionId = window.location.href.split(/requests\/id\//g)[1];
-      if (submissionId.indexOf('&') >= 0) {
-        submissionId = submissionId.split(/&/g)[0];
+  
+  keyLookup(e, fileName) {
+    e.preventDefault();
+    if (this.state.keys[fileName]) {
+      const { dispatch } = this.props;
+      const { requestId } = this.props.match.params;
+      if (requestId !== '' && requestId != undefined && requestId !== null) {
+        const download = new localUpload();
+        const { apiRoot } = _config;
+        dispatch(listFileDownloadsByKey(this.state.keys[fileName], requestId))
+          .then(() => {
+            download.downloadFile(this.state.keys[fileName], `${apiRoot}data/upload/downloadUrl`, loadToken().token).then((resp) => {
+              let error = resp?.data?.error || resp?.error || resp?.data?.[0]?.error
+              if (error) {
+                console.log(`An error has occurred: ${error}.`);
+              }
+            })
+          }
+        );
       }
     }
-    if (submissionId !== '' && submissionId != undefined && submissionId !== null) {
-      dispatch(listFileUploadsBySubmission(submissionId)).then(resp => {
-        /*const bucket = '15df4fda-ed0d-417f-9124-558fb5e5b561';
-        const userId = 'c259a741-1822-48a9-b6c3-9a4ecaac0338';
-         resp = {
-          "id": `${submissionId}`,
-          "type": "REQUEST",
-          "data": [
-            {
-              "key": `${bucket}/${submissionId}/${userId}/DAAC_logo_avatar.png`,
-              "size": 34828,
-              "last_modified": "2023-07-05T19:44:08.000Z",
-              "file_name": "DAAC_logo_avatar.png"
-            },
-            {
-              "key": `${bucket}/${submissionId}/${userId}/home.png`,
-              "size": 435471,
-              "last_modified": "2023-07-20T21:13:30.000Z",
-              "file_name": "home.png"
+  };
+  
+  async getFileList() {
+    const { dispatch } = this.props;
+    const { requestId } = this.props.match.params;
+    if (requestId !== '' && requestId != undefined && requestId !== null) {
+      dispatch(listFileUploadsBySubmission(requestId))
+        .then((resp) => {
+          if (JSON.stringify(resp) === '{}' || JSON.stringify(resp) === '[]' || (resp.data && resp.data.length === 0)) {
+            return
+          }
+          let error = resp?.data?.error || resp?.error || resp?.data?.[0]?.error
+          if (error){
+            if (!error.match(/not authorized/gi) && !error.match(/not implemented/gi)) {
+              const str = `An error has occurred while getting the list of files: ${error}.`;
+              console.log(str)
+              return
+            } else {
+              return
             }
-          ],
-          "config": {
-            "json": true,
-            "resolveWithFullResponse": true,
-            "simple": false,
-            "type": "REQUEST",
-            "method": "GET",
-            "id": `${submissionId}`,
-            "path": `data/upload/list/${submissionId}`,
-            "headers": {
-              "Content-Type": "application/json",
-              "Authorization": `Bearer ${loadToken().token}`
-            },
-            "url": `https://pub.sit.earthdata.nasa.gov/api/data/upload/list/${submissionId}`
           }
-        } */
-        if (resp.error) {
-          const str = `An error has occurred: ${resp.error}.  Please try again later.<br>`
-          this.setState({ files: this.state.files = `${str}` });
-        } else {
-          const dataArr = resp.data
-          for (const ea in dataArr) {
-            const fileName = dataArr[ea].file_name
-            const key = dataArr[ea].key
-            const lastModified = dataArr[ea].last_modified
-            const size = dataArr[ea].size
-            const date = new Date(lastModified).toISOString().split('T')[0];
-            const datetime = date.toLocaleString();
-            const url = resp.config.url
-            const str = `<a target=_blank href="${url}" id="${key}" name="${fileName}" ariaLabel="Download ${key}">${fileName}</a><br>`
-            this.setState({ files: this.state.files += `${str}` });
+
+          const files = resp.data;
+          
+          files.sort(function (a, b) {
+            var keyA = new Date(a.last_modified),
+              keyB = new Date(b.last_modified);
+            if (keyA > keyB) return -1;
+            if (keyA < keyB) return 1;
+            return 0;
+          });
+
+          this.setState({ files: files })
+          
+          const keyDict = {};
+          const html = [];
+
+          for (const ea in files) {
+            const fileName = files[ea].file_name;
+            if (files[ea] === undefined || fileName === undefined) {
+              break
+            }
+            const key = files[ea].key;
+            keyDict[`${fileName}`] = key
           }
-        }
-      });
+          this.setState({ files: files })
+
+          html.map(item =>
+            <span key={item}>{item}</span>  
+          )
+
+          this.setState({ keys: keyDict })
+          this.setState({ saved: html })
+        });
     }
   }
 
-  handleClick (e) {
-    console.log('handle click function')
-    e.preventDefault();
-    if (this.state.hiddenFileInput.current === null || this.state.hiddenFileInput === null) {
-      this.setState({ hiddenFileInput: React.createRef(null) });
+  async componentDidMount() {
+    const { groupId } = this.props.match.params;
+    const { requestId } = this.props.match.params;
+    if ((requestId !== '' && requestId != undefined && requestId !== null) &&
+      (groupId == '' || groupId === undefined || groupId === null)) {
+      await this.getFileList()
+      this.setState({ loaded: true })
     }
+  }
+
+  handleClick(e) {
+    e.preventDefault();
+    this.resetInputWithTimeout(undefined, 0)
     this.state.hiddenFileInput?.current?.click();
   };
 
-  async handleChange (e) {
-    console.log('handle change function')
-    e.preventDefault();
-    this.setState({ statusMsg: 'Uploading' });
-    const file = event.target.files[0];
-    const upload = new localUpload();
-    let submissionId = '';
-    const { apiRoot } = _config;
-    if (window.location.href.indexOf('requests/id') >= 0) {
-      submissionId = window.location.href.split(/requests\/id\//g)[1];
-      if (submissionId.indexOf('&') >= 0) {
-        submissionId = submissionId.split(/&/g)[0];
+  resetInputWithTimeout(msg, timeout) {
+    setTimeout(() => {
+      msg ? this.setState({ statusMsg: msg }) : null
+      if (this.state.hiddenFileInput.current === null || this.state.hiddenFileInput === null) {
+        this.setState({ hiddenFileInput: React.createRef(null) });
       }
+    }, timeout);
+  }
+
+  validateFile(file) {
+    let valid = false;
+    if (file.name.match(/\.([^\.]+)$/) !== null) {
+      var ext = file.name.match(/\.([^\.]+)$/)[1];
+      if (ext.match(/exe/gi)) {
+        this.setState({ statusMsg: 'exe is an invalid file type.' });
+        this.resetInputWithTimeout('Please select a different file.', 2000)
+      } else {
+        valid = true
+      }
+    } else {
+      this.setState({ statusMsg: 'The file must have an extension.' });
+      this.resetInputWithTimeout('Please select a different file.', 2000)
     }
-    if (submissionId !== '' && submissionId != undefined && submissionId !== null) {
-      const payload = {
-        fileObj: file,
-        apiEndpoint: `${apiRoot}data/upload/getPostUrl`,
-        authToken: loadToken().token,
-        submissionId: submissionId
-      }
-      const resp = await upload.uploadFile(payload).then((resp) => {
+    return valid;
+  }
+
+  async handleChange(e) {
+    e.preventDefault();
+    const file = e.target.files[0];
+    if (this.validateFile(file)) {
+      this.setState({ statusMsg: 'Uploading' });
+      const upload = new localUpload();
+      const { requestId } = this.props.match.params;
+      const { groupId } = this.props.match.params;
+      const { apiRoot } = _config;
+      try {
+        let payload = {
+          fileObj: file,
+          authToken: loadToken().token,
+        }
+        let prefix = ''
+        if (requestId !== '' && requestId != undefined && requestId !== null) {
+          payload['apiEndpoint'] = `${apiRoot}data/upload/getPostUrl`;
+          payload['submissionId'] = requestId
+        } else if (groupId !== '' && groupId != undefined && groupId !== null) {
+          if (document.getElementById('prefix') && document.getElementById('prefix') !== null) {
+            prefix = document.getElementById('prefix').value
+          }
+          payload['apiEndpoint'] = `${apiRoot}data/upload/getGroupUploadUrl`;
+          payload['endpointParams'] = {
+            prefix: prefix,
+            group_id: groupId
+          }
+        }
         this.setState({ statusMsg: 'Uploading' });
-        if (resp.error) {
-          console.log(`An error has occured: ${resp.error}.`);
-          setTimeout(() => {
-            this.setState({ statusMsg: 'Select a file' });
-            if (this.state.hiddenFileInput.current === null || this.state.hiddenFileInput === null) {
-              this.setState({ hiddenFileInput: React.createRef(null) });
-            }
-          }, '1000');
+        const resp = await upload.uploadFile(payload)
+        let error = resp?.data?.error || resp?.error || resp?.data?.[0]?.error
+        if (error) {
+          console.log(`An error has occured on uploadFile: ${error}.`);
+          this.resetInputWithTimeout('Select a file', 1000)
         } else {
           this.setState({ statusMsg: 'Upload Complete' });
-          this.updateFileList();
-          setTimeout(() => {
-            this.setState({ statusMsg: 'Select another file' });
-            if (this.state.hiddenFileInput.current === null || this.state.hiddenFileInput === null) {
-              this.setState({ hiddenFileInput: React.createRef(null) });
-            }
-          }, '1000');
+          this.resetInputWithTimeout('Select another file', 1000)
+          if ((requestId !== '' && requestId != undefined && requestId !== null) &&
+            (groupId == '' || groupId === undefined || groupId === null)) {
+            this.getFileList()
+          }
         }
-      })
+      } catch (error) {
+        console.log(`try catch error: ${error.stack}`);
+        this.resetInputWithTimeout('Select a file', 1000)
+      }
     }
   };
 
   render() {
-    console.log('rendering')
+    const tableColumns = [
+      {
+        Header: 'Filename',
+        accessor: row => <><a id={row.file_name} name={row.file_name} aria-label={`Download ${row.file_name}`} onClick={(e) => this.keyLookup(e, row.file_name)}> {row.file_name}</a></>,
+        id: 'file_name'
+      },
+      {
+        Header: 'Size',
+        accessor: row => storage(row.size),
+        id: 'size',
+        width: '100px'
+      },
+      {
+        Header: 'sha256Checksum',
+        accessor: row => row.sha256Checksum,
+        id: 'sha256Checksum'
+      },
+      {
+        Header: 'Last Modified',
+        accessor: row => shortDateShortTimeYearFirstJustValue(row.last_modified),
+        id: 'last_modified'
+      }
+    ];
+    const { requestId } = this.props.match.params;
+    const { groupId } = this.props.match.params;
     return (
       <><br></br>
         <div className='page__component'>
@@ -155,10 +231,14 @@ class UploadOverview extends React.Component {
               Upload Data File
             </h1>
           </div>
-          <div className='indented__details'>
-            <span id='previously-saved' ref={this.state.files}></span>
+          <div className='indented__details' style={{ paddingTop: '1rem' }}>
             <div className='form__textarea'>
-              <br></br><label className='heading--medium' htmlFor='hiddenFileInput' style={{ marginBottom: '1rem' }}>{`${this.state.statusMsg}`}
+              {groupId !== undefined ?
+                <><label htmlFor="prefix" style={{ marginBottom: '1rem', marginTop: '1rem', fontSize: 'unset' }}>Subfolder (If applicable): </label><input id="prefix" name="prefix" style={{ marginBottom: '1rem' }} /></>
+              : null
+              }
+              {this.state.statusMsg === 'Uploading' ? <Loading /> : null}
+              <label htmlFor='hiddenFileInput' style={{ marginBottom: '1rem', fontSize: 'unset' }}>{`${this.state.statusMsg}`}
                 <input
                   onChange={(e) => this.handleChange(e)}
                   type="file"
@@ -169,6 +249,21 @@ class UploadOverview extends React.Component {
               </label>
               <button onClick={(e) => this.handleClick(e)} className={'button button--submit button__animation--md button__arrow button__arrow--md button__animation button__arrow--white'}>Upload File</button>
             </div>
+            {this.state.saved && requestId !== undefined && groupId === undefined
+            ?
+              <><br /><section className = 'page__section'>
+                <div style={{ borderBottom: '1px solid #E2DFDF'}}>
+                <h2 className='heading--medium heading--shared-content with-description'>Files Previously Uploaded</h2>
+                </div>
+                <Table
+                  data={this.state.files}
+                  dispatch={this.props.dispatch}
+                  tableColumns={tableColumns}
+                />
+              </section></>
+              : null }
+            {!this.state.loaded && groupId === undefined ? <Loading /> : null}
+            <span>{this.state.saved ? this.state.saved : null}</span>
           </div>
         </div></>
     );
@@ -178,10 +273,12 @@ class UploadOverview extends React.Component {
 UploadOverview.propTypes = {
   stats: PropTypes.object,
   dispatch: PropTypes.func,
-  config: PropTypes.object
+  config: PropTypes.object,
+  logs: PropTypes.object,
 };
 
 export default withRouter(connect(state => ({
   stats: state.stats,
-  config: state.config
+  config: state.config,
+  logs: state.logs
 }))(UploadOverview));