+++ app/src/js/components/Groups/group.js
@@ -3,42 +3,19 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import {
-  withRouter,
-  // Link
+  withRouter
 } from 'react-router-dom';
 import {
-  interval,
   getGroup,
-  deleteGroup
 } from '../../actions';
-import { get } from 'object-path';
-import {
-// fromNow,
-// lastUpdated,
-// deleteText
-} from '../../utils/format';
 import Loading from '../LoadingIndicator/loading-indicator';
-// import AsyncCommands from '../DropDown/dropdown-async-command';
-import ErrorReport from '../Errors/report';
+import { groupPrivileges } from '../../utils/privileges';
+import UploadOverview from '../DataUpload/overview';
 import Metadata from '../Table/Metadata';
+import { strings } from '../locale';
+import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
 import _config from '../../config';
 
-const { updateInterval } = _config;
-
-/* const permissionTableColumns = [
-  {
-    Header: 'Table Name',
-    accessor: row => row
-  }
-];
-
-const subscriptionTableColumns = [
-  {
-    Header: 'Table Name',
-    accessor: row => row,
-  }
-];
-*/
 const metaAccessors = [
   {
     label: 'Short Name',
@@ -55,30 +32,20 @@ const metaAccessors = [
 ];
 
 class GroupOverview extends React.Component {
-  constructor () {
-    super();
-    this.reload = this.reload.bind(this);
+  constructor (props) {
+    super(props);
+    this.state = {};
     this.navigateBack = this.navigateBack.bind(this);
-    this.delete = this.delete.bind(this);
-    this.errors = this.errors.bind(this);
   }
 
   componentDidMount () {
+    const { dispatch } = this.props;
     const { groupId } = this.props.match.params;
-    const immediate = !this.props.groups.map[groupId];
-    this.reload(immediate);
+    dispatch(getGroup(groupId));
   }
 
   componentWillUnmount () {
-    if (this.cancelInterval) { this.cancelInterval(); }
-  }
 
-  reload (immediate, timeout) {
-    timeout = timeout || updateInterval;
-    const groupId = this.props.match.params.groupId;
-    const { dispatch } = this.props;
-    if (this.cancelInterval) { this.cancelInterval(); }
-    this.cancelInterval = interval(() => dispatch(getGroup(groupId)), timeout, immediate);
   }
 
   navigateBack () {
@@ -86,83 +53,52 @@ class GroupOverview extends React.Component {
     history.push('/groups');
   }
 
-  delete () {
-    const { groupId } = this.props.match.params;
-    const group = this.props.groups.map[groupId].data;
-    if (!group.published) {
-      this.props.dispatch(deleteGroup(groupId));
-    }
-  }
-
-  errors () {
-    const groupId = this.props.match.params.groupId;
-    return [
-      get(this.props.groups.map, [groupId, 'error']),
-      get(this.props.groups.deleted, [groupId, 'error'])
-    ].filter(Boolean);
-  }
-
   render () {
     const groupId = this.props.match.params.groupId;
     const record = this.props.groups.map[groupId];
-
+    let { canUpload } = groupPrivileges(this.props.privileges);
     if (!record || (record.inflight && !record.data)) {
       return <Loading />;
-    } else if (record.error) {
-      return <ErrorReport report={record.error} truncate={true} />;
     }
     const group = record.data;
-
-    const errors = this.errors();
-
-    // const deleteStatus = get(this.props.groups.deleted, [groupId, 'status']);
-    /* const dropdownConfig = [{
-      text: 'Delete',
-      action: this.delete,
-      disabled: group.published,
-      status: deleteStatus,
-      success: this.navigateBack,
-      confirmAction: true,
-      confirmText: deleteText(groupId)
-    }]; */
-
+    if (group.short_name.match(/unassigned/)) {
+      canUpload = false;
+    }
+    const breadcrumbConfig = [
+      {
+        label: 'Dashboard Home',
+        href: '/'
+      },
+      {
+        label: 'Groups',
+        href: '/groups'
+      },
+      {
+        label: groupId,
+        active: true
+      }
+    ];
     return (
       <div className='page__component'>
+        <section className='page__section page__section__controls'>
+          <Breadcrumbs config={breadcrumbConfig} />
+        </section>
         <section className='page__section page__section__header-wrapper'>
-          <h1 className='heading--large heading--shared-content with-description'>{groupId}</h1>
-          {/* <AsyncCommands config={dropdownConfig} />
-          <Link
-            className='button button--small button--green button--edit form-group__element--right'
-            to={'/groups/edit/' + groupId}>Edit</Link> */}
-        </section>
-
-        <section className='page__section'>
-          {errors.length ? <ErrorReport report={errors} truncate={true} /> : null}
-          <div className='heading__wrapper--border'>
-            <h2 className='heading--medium with-description'>Group Overview</h2>
+          <div className='page__section__header'>
+            <h1 className='heading--large heading--shared-content with-description '>
+              Group Overview
+            </h1>
           </div>
-          <Metadata data={group} accessors={metaAccessors} />
         </section>
-
-        {/* <section className='page__section'>
-          <div className='heading__wrapper--border'>
-            <h2 className='heading--medium heading--shared-content with-description'>Permissions</h2>
-          </div>
-          <Table
-            data={group.permissions}
-            tableColumns={permissionTableColumns}
-          />
-        </section>
-
         <section className='page__section'>
           <div className='heading__wrapper--border'>
-            <h2 className='heading--medium heading--shared-content with-description'>Subscriptions</h2>
+            <h1 className='heading--small' aria-labelledby={strings.group_overview}>
+              {strings.group_overview}
+            </h1>
           </div>
-          <Table
-            data={group.subscriptions}
-            tableColumns={subscriptionTableColumns}
-          />
-        </section> */}
+          <div className='indented__details'><Metadata data={group} accessors={metaAccessors} /></div>
+          {record.data && canUpload && (_config.fileUploadDefault === 'true') ? <UploadOverview /> : null }
+        </section>
       </div>
     );
   }
@@ -173,12 +109,13 @@ GroupOverview.propTypes = {
   dispatch: PropTypes.func,
   groups: PropTypes.object,
   logs: PropTypes.object,
-  history: PropTypes.object
+  history: PropTypes.object,
+  privileges: PropTypes.object,
+  roles: PropTypes.array
 };
-
-GroupOverview.displayName = 'GroupElem';
 
 export default withRouter(connect(state => ({
   groups: state.groups,
+  privileges: state.api.tokens.privileges,
   logs: state.logs
 }))(GroupOverview));