+++ app/src/js/utils/auth.js
@@ -2,7 +2,7 @@
 import { window } from './browser';
 
 export const saveToken = function ({ token, user }) {
-  if (window.localStorage && typeof window.localStorage.setItem === 'function') {
+  if (window.localStorage && typeof window.localStorage.setItem === 'function' && token !== undefined && token !== null && token !== 'undefined') {
     window.localStorage.setItem('auth-token', token);
     window.localStorage.setItem('auth-user', JSON.stringify(user));
   }