+++ CHANGELOG.md
@@ -10,6 +10,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- Unreleased changes can be added here. -->
 
+## 1.0.19
+
+- Added config variable for file upload
+- Updated css for some font discrepancies
+- Added in group upload functionality
+- Added group privileges
+- Added downloading of files to Data Upload.
+- Added getting list of files to Data Upload.
+- Added missing types for some components.
+- Added auth token refresh at start of file upload to reduce issue with timeout during file upload.
+
+## 1.0.18
+
+- Add token refresh before upload
+- Add import
+- Update edit perms to allow daac staff to edit
+
 ## 1.0.17
 
 - Update upload formatting add file listing